--- server/templates/place.html
@@ -1,102 +1,0 @@
-{#
-Copyright 2020 Google LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-#}
-{%- extends BASE_HTML -%}
-
-{% from 'macros/icons.html' import inline_svg %}
-
-{% set main_id = 'dc-places' %}
-{% set page_id = 'page-dc-places' %}
-{% if category == '' %}
-  {% set title = place_name + ' - ' + _('Place Explorer') %}
-  {% set description = _('Statistics about economics, health, equity, crime, education, demographics, housing, and environment in {place_name}.'.format(place_name=place_name)) %}
-{% else %}
-  {% set title = place_name + ' - '+ _('CHART_TITLE-CHART_CATEGORY-{category}'.format(category=category)) + ' - ' + _('Place Explorer') %}
-  {% set description = _('Statistics about {category} in {place_name}.'.format(category=category, place_name=place_name)) %}
-{% endif %}
-{% set place_category = category %}
-{% set new_place_page = false %}
-
-{% block head %}
-<link rel="stylesheet" href={{url_for('static', filename='css/place_page.min.css' , t=config['VERSION'])}}>
-<link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" />
-
-{% if locale == "en" %}
-<meta name="description" content="{{ place_summary or description }}" />
-{% endif %}
-
-{% if block_indexing %}
-<meta name="robots" content="noindex" />
-{% endif %}
-
-{% endblock %}
-
-{% block content %}
-<div id="body" class="container-fluid">
-  <div id="nl-search-bar"></div>
-  <div id="body-row" class="row">
-    <div id="sidebar-outer" class="col-md-3x col-xl-2 order-last order-xl-0">
-      <div id="sidebar-top-spacer" class="d-none d-xl-block"></div>
-      <div id="sidebar-region">
-        <div id="menu"></div>
-        <div id="child-place"></div>
-      </div>
-    </div>
-    <div id="main-place" class="col-md-12 col-xl-10">
-      <div id="title" data-dcid="{{ place_dcid }}">
-        <div id="place-heading-container">
-          <div id="place-heading">
-            <h1 id="place-name" data-pn="{{ place_name }}">{{ place_name }}</h1>
-            <h3 id="place-type" data-pt="{{ place_type }}">{{ place_type_with_parent_places_links | safe }}</h3>
-            <div id="place-highlight"></div>
-          </div>
-          <div class="dcid-and-knowledge-graph">
-            dcid:&nbsp;<a href="/browser/{{ place_dcid }}">{{ place_dcid }}</a>
-          </div>
-        </div>
-      </div>
-      <div id="subtitle" class="row"></div>
-      <div id="place-summary" class="row-col place-summary-container">{{ place_summary }}</div>
-      <div id="main-pane" class="row"></div>
-      <div id="page-loading" class="mt-4">
-        {{ inline_svg('progress_activity') }}
-        {# TRANSLATORS: A message shown on the page while the content is loading. #}
-        <p>{% trans %}Loading{% endtrans %}</p>
-      </div>
-    </div>
-  </div>
-</div>
-<div id="modal"></div>
-{#- Add an SVG filter to lighten bars for comparison places -#}
-<style>
-  g rect.g-bar[data-dcid="{{ place_dcid }}"] {
-    filter: url(#lighten);
-  }
-</style>
-<svg width=0 height=0>
-  <filter id="lighten">
-    <feColorMatrix type="matrix" values="1.25 0    0    0    0
-                                         0    1.25 0    0    0
-                                         0    0    1.25 0    0
-                                         0    0    0    1    0" />
-  </filter>
-</svg>
-{% endblock %}
-
-{% block footer %}
-<script src={{url_for('static', filename='place.js' , t=config['VERSION'])}}></script>
-<script src="https://maps.googleapis.com/maps/api/js?key={{maps_api_key}}&libraries=places&language={{locale}}&loading=async" async
-  defer></script>
-{% endblock %}
