+++ server/tests/routes/place_test.py
@@ -154,144 +154,6 @@ class TestPlacePage(unittest.TestCase):
       assert '/place?dcid=geoId/06&utm_medium=explore&mprop=count&popt=Person&hl=fr' in request.url
       assert response.status_code == 200
 
-
-class TestPlacePageHeaders(unittest.TestCase):
-
-  @patch('server.routes.shared_api.place.get_i18n_name')
-  @patch('server.routes.shared_api.place.api_place_type')
-  @patch('server.routes.shared_api.place.get_place_type_i18n_name')
-  @patch('server.routes.shared_api.place.parent_places')
-  def test_place_page_canonical_header(self, mock_parent_places,
-                                       mock_get_place_type_i18n_name,
-                                       mock_api_place_type, mock_get_i18n_name):
-    mock_feature_flags(app, ['dev_place_ga'], False)
-    mock_parent_places.return_value = {
-        'geoId/06': [{
-            'dcid': 'country/USA',
-            'type': 'Country',
-            'name': 'United States'
-        }]
-    }
-    mock_get_place_type_i18n_name.return_value = 'State'
-    mock_get_i18n_name.return_value = {'geoId/06': 'California'}
-    mock_api_place_type.return_value = 'State'
-
-    # Test main page gives canonical without query parameters
-    response = app.test_client().get('/place/geoId/06', follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test "Overview" page gives canonical without query parameters
-    response = app.test_client().get('/place/geoId/06', follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test category page gives a canonical with category included
-    response = app.test_client().get('/place/geoId/06?category=Health',
-                                     follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06?category=Health>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test English locale page gives canonical without locale included (default)
-    response = app.test_client().get('/place/geoId/06?hl=en',
-                                     follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test localized main page gives a canonical with locale included
-    response = app.test_client().get('/place/geoId/06?hl=ru',
-                                     follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06?hl=ru>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test localized category page gives a canonical with locale included.
-    response = app.test_client().get('/place/geoId/06?category=Health&hl=ru',
-                                     follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06?category=Health&hl=ru>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test bad category input gives a canonical without category
-    response = app.test_client().get('/place/geoId/06?category=foobar',
-                                     follow_redirects=True)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06>; rel="canonical"' in response.headers.get(
-        'Link')
-
-    # Test bad locale input gives a canonical without locale
-    response = app.test_client().get('/place/geoId/06?hl=foobar',
-                                     follow_redirects=True)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06>; rel="canonical"' in response.headers.get(
-        'Link')
-
-  @patch('server.routes.shared_api.place.get_i18n_name')
-  @patch('server.routes.shared_api.place.api_place_type')
-  @patch('server.routes.shared_api.place.get_place_type_i18n_name')
-  @patch('server.routes.shared_api.place.parent_places')
-  def test_place_page_alternate_header(self, mock_parent_places,
-                                       mock_get_place_type_i18n_name,
-                                       mock_api_place_type, mock_get_i18n_name):
-    mock_feature_flags(app, ['dev_place_ga'], False)
-    mock_parent_places.return_value = {
-        'geoId/06': [{
-            'dcid': 'country/USA',
-            'type': 'Country',
-            'name': 'United States'
-        }]
-    }
-    mock_get_place_type_i18n_name.return_value = 'State'
-    mock_get_i18n_name.return_value = {'geoId/06': 'California'}
-    mock_api_place_type.return_value = 'State'
-
-    # Test available languages listed as alternates for main page
-    response = app.test_client().get('/place/geoId/06', follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06>; rel="alternate"; hreflang="en"' in response.headers.get(
-        'Link')
-    assert '<https://datacommons.org/place/geoId/06?hl=ru>; rel="alternate"; hreflang="ru"' in response.headers.get(
-        'Link')
-    assert '<https://datacommons.org/place/geoId/06>; rel="alternate"; hreflang="x-default"' in response.headers.get(
-        'Link')
-
-    # Test available languages listed as alternates for category page
-    response = app.test_client().get('/place/geoId/06?category=Health',
-                                     follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06?category=Health>; rel="alternate"; hreflang="en"' in response.headers.get(
-        'Link')
-    assert '<https://datacommons.org/place/geoId/06?category=Health&hl=ru>; rel="alternate"; hreflang="ru"' in response.headers.get(
-        'Link')
-    assert '<https://datacommons.org/place/geoId/06?category=Health>; rel="alternate"; hreflang="x-default"' in response.headers.get(
-        'Link')
-
-    # Test available languages listed as alternates for localized page
-    response = app.test_client().get('/place/geoId/06?category=Health&hl=ru',
-                                     follow_redirects=False)
-    assert response.status_code == 200
-    assert 'Link' in response.headers
-    assert '<https://datacommons.org/place/geoId/06?category=Health>; rel="alternate"; hreflang="en"' in response.headers.get(
-        'Link')
-    assert '<https://datacommons.org/place/geoId/06?category=Health&hl=ru>; rel="alternate"; hreflang="ru"' in response.headers.get(
-        'Link')
-    assert '<https://datacommons.org/place/geoId/06?category=Health>; rel="alternate"; hreflang="x-default"' in response.headers.get(
-        'Link')
-
   @patch('server.routes.shared_api.place.get_i18n_name')
   @patch('server.routes.shared_api.place.api_place_type')
   @patch('server.routes.shared_api.place.get_place_type_i18n_name')
