+++ static/js/chart/types.ts
@@ -30,14 +30,6 @@ export interface Series {
   [key: string]: number;
 }
 
-// TrendData represents a set of time series to be used in a multi-line chart.
-export interface TrendData {
-  series: { [key: string]: Series };
-  sources: string[];
-  exploreUrl?: string;
-  statsVars?: string[];
-}
-
 export interface SnapshotData {
   date: string;
   data: {
@@ -50,60 +42,8 @@ export interface SnapshotData {
   statsVars?: string[];
 }
 
-export interface ChartBlockData {
-  title: string;
-  statsVars: string[];
-  denominator?: string[];
-  unit: string;
-  trend: TrendData;
-  parent: SnapshotData;
-  similar: SnapshotData;
-  nearby: SnapshotData;
-  child: SnapshotData;
-  relatedChart: {
-    title?: string;
-    scale: boolean;
-    scaling?: number;
-    unit?: string;
-  };
-  scaling: number;
-  isChoropleth?: boolean;
-  isRankingChart?: boolean;
-}
-
-export interface PageChart {
-  [key: string]: {
-    [key: string]: ChartBlockData[];
-  };
-}
-
 export interface PageHighlight {
   [title: string]: SnapshotData;
-}
-
-export interface PageData {
-  pageChart: PageChart;
-  allChildPlaces: {
-    string: string[];
-  };
-  childPlacesType: string;
-  childPlaces: string[];
-  parentPlaces: string[];
-  similarPlaces: string[];
-  nearbyPlaces: string[];
-  categories: { string: string };
-  names: { string: string };
-  highlight: PageHighlight;
-}
-
-export interface ChoroplethDataGroup {
-  date: string;
-  data: {
-    [placeDcid: string]: number;
-  };
-  numDataPoints: number;
-  exploreUrl: string;
-  sources: string[];
 }
 
 export interface GeoJsonFeatureProperties {
