+++ static/webpack.config.js
@@ -37,8 +37,14 @@ const config = {
       __dirname + "/css/tools/stat_var.scss",
     ],
     dev: [__dirname + "/js/dev.ts", __dirname + "/css/dev.scss"],
-    datagemma: [__dirname + "/js/apps/datagemma/main.ts", __dirname + "/css/datagemma.scss"],
-    biomed_nl: [__dirname + "/js/apps/biomed_nl/main.ts", __dirname + "/css/biomed_nl.scss"],
+    datagemma: [
+      __dirname + "/js/apps/datagemma/main.ts",
+      __dirname + "/css/datagemma.scss",
+    ],
+    biomed_nl: [
+      __dirname + "/js/apps/biomed_nl/main.ts",
+      __dirname + "/css/biomed_nl.scss",
+    ],
     diff: [__dirname + "/js/apps/diff/main.ts", __dirname + "/css/diff.scss"],
     timeline: [
       __dirname + "/js/tools/timeline/timeline.ts",
@@ -48,10 +54,6 @@ const config = {
     mcf_playground: __dirname + "/js/mcf_playground.js",
     queryStore: path.resolve(__dirname, "js/shared/stores/query_store.ts"),
     base: [__dirname + "/js/apps/base/main.ts", __dirname + "/css/core.scss"],
-    place: [
-      __dirname + "/js/place/place.ts",
-      __dirname + "/css/place/place_page.scss",
-    ],
     place_landing: [
       __dirname + "/js/place/place_landing.ts",
       __dirname + "/css/place/place_landing.scss",
@@ -138,9 +140,7 @@ const config = {
       __dirname + "/js/apps/build/main.ts",
       __dirname + "/css/build.scss",
     ],
-    data: [
-      __dirname + "/js/apps/data/main.ts",
-    ],
+    data: [__dirname + "/js/apps/data/main.ts"],
     data_overview: [
       __dirname + "/js/apps/data_overview/main.ts",
       __dirname + "/css/data_overview.scss",
